api/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
hindsight-client
